# packages/specs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: App API
  version: 0.15.0
  description: |
    Spec-first HTTP API. Paths include the versioned prefix (`/api/v1`) so
    generated clients can be used without additional base-URL rewriting.

    **Authentication endpoints** live under `/api/v1/auth/*` and follow the
    [Better Auth protocol](https://www.better-auth.com/docs). They are
    deliberately not enumerated here: Better Auth owns that wire contract and
    evolves it independently of our OpenAPI surface. The `cookieAuth` and
    `bearerAuth` security schemes documented below describe how the *rest* of
    this API accepts credentials issued by that namespace — session cookies
    (web) and bearer tokens (mobile, via the `bearer` plugin).
  contact:
    name: App Team
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.example.com
    description: Production (placeholder)
...
paths:
...
  /api/v1/courses:
    get:
      operationId: listCourses
      summary: List courses (with filtering and sort)
      description: |
        Returns courses the requester can see. Non-admins see only courses inside
        libraries they have a READ AccessGrant for; admins see all.

        The `status` and `sort` query params back the Browse page (E14-F01-S02).
        `status` filters by per-user progress derived from the
        CourseProgressReadModel projection. `sort` is server-applied so the
        SPA never needs to re-sort the response.
      tags: [Catalog]
      security:
        - bearerAuth: []
      parameters:
        - name: libraryId
          in: query
          required: false
          description: Filter to a single library; omit for everything visible.
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: |
            Filter by per-user progress state.
              - `all` (default) — no filter.
              - `not-started` — `progress.percent == 0` and no `lessonsCompleted`.
              - `in-progress` — `0 < progress.percent < 100`.
              - `completed`   — `progress.percent == 100`.
          schema:
            type: string
            enum: [all, not-started, in-progress, completed]
            default: all
        - name: sort
          in: query
          required: false
          description: |
            Server-side sort. `recently-watched` (default) uses
            `updatedAt` as a proxy for last activity until a dedicated
            `lastViewedAt` field is added. `newest` is `createdAt` desc.
            `alphabetical` is title asc, locale-insensitive.
          schema:
            type: string
            enum: [recently-watched, newest, alphabetical]
            default: recently-watched
      responses:
        "200":
          description: Course list returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseListDto"
              examples:
                list:
                  value:
                    items:
                      - id: clxvcrs0000000000000000001
                        libraryId: clxvp1234567890abcdefghij
                        slug: pragmatic-clean-architecture
                        title: Pragmatic Clean Architecture
                        description: Milan Jovanović's deep-dive into Clean Architecture with .NET.
                        sections:
                          - id: clxvsec0000000000000000001
                            position: 1
                            title: Introduction
                          - id: clxvsec0000000000000000002
                            position: 2
                            title: Domain Layer
                        progress:
                          percent: 0
                          lessonsCompleted: 0
                          lessonsTotal: 24
                        createdAt: "2026-04-25T09:00:00Z"
                        updatedAt: "2026-04-25T09:00:00Z"
                      - id: clxvcrs0000000000000000002
                        libraryId: clxvp1234567890abcdefghij
                        slug: ddd-by-eric-evans
                        title: Domain-Driven Design by Eric Evans
                        sections:
                          - id: clxvsec0000000000000000003
                            position: 1
                            title: Putting the Domain Model to Work
                          - id: clxvsec0000000000000000004
                            position: 2
                            title: Building Blocks of a Model-Driven Design
                        progress:
                          percent: 0
                          lessonsCompleted: 0
                          lessonsTotal: 18
                        createdAt: "2026-04-25T09:05:00Z"
                        updatedAt: "2026-04-25T09:05:00Z"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
  /api/v1/courses/{id}:
    get:
      operationId: getCourse
      summary: Get a single course
      description: Returns the full CourseDto for one course by its server-generated cuid. Non-admins must hold a READ AccessGrant on the course's library.
      tags: [Catalog]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Server-generated cuid identifying the course.
          schema:
            type: string
      responses:
        "200":
          description: Course found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDto"
              examples:
                found:
                  value:
                    id: clxvcrs0000000000000000001
                    libraryId: clxvp1234567890abcdefghij
                    slug: pragmatic-clean-architecture
                    title: Pragmatic Clean Architecture
                    description: Milan Jovanović's deep-dive into Clean Architecture with .NET.
                    sections:
                      - id: clxvsec0000000000000000001
                        position: 1
                        title: Introduction
                      - id: clxvsec0000000000000000002
                        position: 2
                        title: Domain Layer
                    progress:
                      percent: 0
                      lessonsCompleted: 0
                      lessonsTotal: 24
                    createdAt: "2026-04-25T09:00:00Z"
                    updatedAt: "2026-04-25T09:00:00Z"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
        "403":
          description: Caller does not have a READ grant for the course's library
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                forbidden:
                  value:
                    type: about:blank
                    title: Forbidden
                    status: 403
                    detail: You do not have access to this course.
        "404":
          description: Course not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: about:blank
                    title: Not Found
                    status: 404
                    detail: No course with that id exists.

    patch:
      operationId: updateCourse
      summary: Update course metadata
      description: |
        Admin-only. Updates any combination of title / description / slug.
        Slug must be unique within the same library.
        At least one of `title`, `description`, or `slug` must be present
        (server-side validation rule — OpenAPI cannot express "at-least-one" natively).
      tags: [Catalog]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Server-generated cuid identifying the course to update.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseRequest"
            examples:
              renameAndReslug:
                summary: Rename title and update slug
                value:
                  title: Pragmatic Clean Architecture (2nd ed.)
                  slug: pragmatic-clean-architecture-2nd-ed
      responses:
        "200":
          description: Course updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDto"
              examples:
                updated:
                  value:
                    id: clxvcrs0000000000000000001
                    libraryId: clxvp1234567890abcdefghij
                    slug: pragmatic-clean-architecture-2nd-ed
                    title: Pragmatic Clean Architecture (2nd ed.)
                    description: Milan Jovanović's deep-dive into Clean Architecture with .NET.
                    sections:
                      - id: clxvsec0000000000000000001
                        position: 1
                        title: Introduction
                      - id: clxvsec0000000000000000002
                        position: 2
                        title: Domain Layer
                    progress:
                      percent: 0
                      lessonsCompleted: 0
                      lessonsTotal: 24
                    createdAt: "2026-04-25T09:00:00Z"
                    updatedAt: "2026-04-25T10:15:00Z"
        "400":
          description: Validation error — malformed fields or no fields provided
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                badInput:
                  value:
                    type: about:blank
                    title: Bad Request
                    status: 400
                    detail: At least one of title, description, or slug must be present.
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
        "403":
          description: Caller does not have the admin role
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                forbidden:
                  value:
                    type: about:blank
                    title: Forbidden
                    status: 403
                    detail: Only admins may update course metadata.
        "404":
          description: Course not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                notFound:
                  value:
                    type: about:blank
                    title: Not Found
                    status: 404
                    detail: No course with that id exists.
        "409":
          description: A course with the same slug already exists in this library
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                conflict:
                  value:
                    type: about:blank
                    title: Conflict
                    status: 409
                    detail: A course with slug 'pragmatic-clean-architecture-2nd-ed' already exists in this library.
  /api/v1/courses/{id}/outline:
    get:
      operationId: getCourseOutline
      summary: Full course outline — sections, lessons (lite), and aggregated materials
      description: |
        Single round-trip endpoint feeding the Course detail page. Returns
        the course summary, every section with its lesson list (lightweight:
        title, duration, hasMaterials, per-user progress), and a flat list of
        course-level materials aggregated across all lessons. The dedicated
        outline avoids N+1 page fetches and never returns full LessonDtos
        (which would inflate the payload with subtitle tracks the page
        does not render).

        Reads `Course` + `Section` + `Lesson` + `Material` + `LessonProgress`
        (filtered to the requester) + `CourseProgressReadModel` (for the
        aggregate progress percent).
      tags: [Catalog]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Server-generated cuid identifying the course.
          schema:
            type: string
      responses:
        "200":
          description: Course outline returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseOutlineDto"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
        "403":
          description: Caller does not have a READ grant for the course's library
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Course not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/courses/{id}/mark-complete:
    post:
      operationId: markCourseComplete
      summary: Mark every lesson in the course as completed for the requester
      description: |
        Bulk-marks every lesson in the course as completed for the
        requester. Idempotent — a second call is a no-op. Returns the
        refreshed `CourseOutlineDto` so the caller does not have to issue
        a separate GET.

        Implementation note: the handler upserts `LessonProgress` rows
        with `completed: true`, `completedAt: now`, and
        `positionSeconds: durationSeconds`. `CourseProgressReadModel` is
        kept in sync via the `LessonCompleted` event handler.
      tags: [Learning]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Server-generated cuid identifying the course.
          schema:
            type: string
      responses:
        "200":
          description: Every lesson marked complete; refreshed outline returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseOutlineDto"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Caller does not have a READ grant for the course's library
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Course not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/courses/{id}/reset-progress:
    post:
      operationId: resetCourseProgress
      summary: Clear every progress row in the course for the requester
      description: |
        Deletes every `LessonProgress` row for (requester, course).
        Idempotent — a second call is a no-op. Returns the refreshed
        `CourseOutlineDto` so the caller does not have to issue a
        separate GET.

        `CourseProgressReadModel` is kept in sync via the
        `LessonProgressReset` event handler (or rebuilt directly when
        no events are emitted on delete).
      tags: [Learning]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Server-generated cuid identifying the course.
          schema:
            type: string
      responses:
        "200":
          description: Progress cleared; refreshed outline returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseOutlineDto"
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Caller does not have a READ grant for the course's library
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Course not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Better Auth bearer token
  schemas:
...
    CourseDto:
      type: object
      required: [id, libraryId, slug, title, sections, progress, createdAt, updatedAt]
      description: Full representation of a Course aggregate.
      properties:
        id:
          type: string
          description: Server-generated cuid identifying this course.
        libraryId:
          type: string
          description: cuid of the library this course belongs to.
        slug:
          $ref: "#/components/schemas/CourseSlug"
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 4000
        sections:
          type: array
          items:
            $ref: "#/components/schemas/SectionDto"
          description: Sections sorted ascending by position.
        progress:
          $ref: "#/components/schemas/CourseProgress"
        instructors:
          type: array
          items:
            $ref: "#/components/schemas/InstructorRef"
          default: []
          description: Instructors associated with this course. Empty array when none linked.
        studios:
          type: array
          items:
            $ref: "#/components/schemas/StudioRef"
          default: []
          description: Studios associated with this course. Empty array when none linked.
        tags:
          type: array
          items:
            $ref: "#/components/schemas/TagRef"
          default: []
          description: Tags associated with this course. Empty array when none linked.
        level:
          oneOf:
            - $ref: "#/components/schemas/CourseLevel"
            - type: "null"
          description: Difficulty level of the course, or null when not set.
        language:
          oneOf:
            - type: string
            - type: "null"
          description: >-
            BCP-47 language tag for the course primary language (e.g. `en`,
            `en-US`). Validated server-side; null when not set.
        releaseDate:
          oneOf:
            - type: string
              format: date
            - type: "null"
          description: ISO-8601 date when the course was originally released. Null when not set.
        posterUrl:
          oneOf:
            - type: string
              format: uri
            - type: "null"
          description: URL of the course poster image. Null when not set.
        ratingAverage:
          oneOf:
            - type: number
              multipleOf: 0.01
              minimum: 0
              maximum: 5
            - type: "null"
          description: Average rating on a 0–5 scale (two decimal places). Null when not set.
        ratingCount:
          oneOf:
            - type: integer
              minimum: 0
            - type: "null"
          description: Number of ratings contributing to `ratingAverage`. Null when not set.
        externalIds:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdRef"
          default: []
          description: External system references for this course. Empty array when none.
        sourceUpdatedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
          description: Timestamp from the upstream source when this course was last updated. Null when not set.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        id: clxvcrs0000000000000000001
        libraryId: clxvp1234567890abcdefghij
        slug: pragmatic-clean-architecture
        title: Pragmatic Clean Architecture
        description: Milan Jovanović's deep-dive into Clean Architecture with .NET.
        sections:
          - id: clxvsec0000000000000000001
            position: 1
            title: Introduction
          - id: clxvsec0000000000000000002
            position: 2
            title: Domain Layer
        progress:
          percent: 0
          lessonsCompleted: 0
          lessonsTotal: 24
        instructors: []
        studios: []
        tags: []
        level: null
        language: null
        releaseDate: null
        posterUrl: null
        ratingAverage: null
        ratingCount: null
        externalIds: []
        sourceUpdatedAt: null
        createdAt: "2026-04-25T09:00:00Z"
        updatedAt: "2026-04-25T09:00:00Z"
    CourseListDto:
      type: object
      required: [items]
      description: List of courses visible to the requester.
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CourseDto"
      example:
        items:
          - id: clxvcrs0000000000000000001
            libraryId: clxvp1234567890abcdefghij
            slug: pragmatic-clean-architecture
            title: Pragmatic Clean Architecture
            description: Milan Jovanović's deep-dive into Clean Architecture with .NET.
            sections:
              - id: clxvsec0000000000000000001
                position: 1
                title: Introduction
              - id: clxvsec0000000000000000002
                position: 2
                title: Domain Layer
            progress:
              percent: 0
              lessonsCompleted: 0
              lessonsTotal: 24
            createdAt: "2026-04-25T09:00:00Z"
            updatedAt: "2026-04-25T09:00:00Z"
          - id: clxvcrs0000000000000000002
            libraryId: clxvp1234567890abcdefghij
            slug: ddd-by-eric-evans
            title: Domain-Driven Design by Eric Evans
            sections:
              - id: clxvsec0000000000000000003
                position: 1
                title: Putting the Domain Model to Work
              - id: clxvsec0000000000000000004
                position: 2
                title: Building Blocks of a Model-Driven Design
            progress:
              percent: 0
              lessonsCompleted: 0
              lessonsTotal: 18
            createdAt: "2026-04-25T09:05:00Z"
            updatedAt: "2026-04-25T09:05:00Z"
...
    CourseSlug:
      type: string
      pattern: "^[a-z0-9](?:[a-z0-9-]{0,98}[a-z0-9])?$"
      description: >-
        URL-safe slug. 1–100 chars, lowercase ASCII letters, digits, and hyphens;
        cannot start or end with a hyphen. Unique within a library.
      example: pragmatic-clean-architecture
...
    Problem:
      type: object
      description: RFC 9457 problem details
      required: [type, title, status]
      properties:
        type:
          type: ["string", "null"]
          format: uri
          default: about:blank
        title:
          type: ["string", "null"]
        status:
          type: integer
          format: int32
        detail:
          type: ["string", "null"]
        instance:
          type: ["string", "null"]
        code:
          type: ["string", "null"]
...
    UpdateCourseRequest:
      type: object
      description: |
        Payload for updating course metadata. All fields are optional, but at
        least one of `title`, `description`, `slug`, `instructorIds`,
        `studioIds`, `tagIds`, `posterUrl`, `level`, `language`,
        `releaseDate`, `sourceUpdatedAt`, `ratingAverage`, `ratingCount`, or
        `externalIds` must be present (server-side validation rule — OpenAPI
        does not have a native "at-least-one" constraint).

        **Set-replace semantics for relation arrays:** `null` means "leave the
        existing set alone"; `[]` (empty array) means "remove all links";
        a non-empty array replaces the full set with exactly the listed ids.
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 4000
        slug:
          $ref: "#/components/schemas/CourseSlug"
        instructorIds:
          oneOf:
            - type: array
              items:
                type: string
            - type: "null"
          description: >-
            Set-replace list of instructor cuids. `null` = leave unchanged;
            `[]` = remove all instructor links; `[id1, id2]` = replace with
            exactly these instructors.
        studioIds:
          oneOf:
            - type: array
              items:
                type: string
            - type: "null"
          description: >-
            Set-replace list of studio cuids. Same null/empty/array semantics
            as `instructorIds`.
        tagIds:
          oneOf:
            - type: array
              items:
                type: string
            - type: "null"
          description: >-
            Set-replace list of tag cuids. Same null/empty/array semantics as
            `instructorIds`.
        level:
          oneOf:
            - $ref: "#/components/schemas/CourseLevel"
            - type: "null"
          description: Difficulty level. Pass `null` to clear.
        language:
          oneOf:
            - type: string
            - type: "null"
          description: BCP-47 language tag (e.g. `en`, `en-US`). Pass `null` to clear.
        releaseDate:
          oneOf:
            - type: string
              format: date
            - type: "null"
          description: ISO-8601 date of original release. Pass `null` to clear.
        posterUrl:
          oneOf:
            - type: string
              format: uri
            - type: "null"
          description: Course poster image URL. Pass `null` to clear.
        ratingAverage:
          oneOf:
            - type: number
              multipleOf: 0.01
              minimum: 0
              maximum: 5
            - type: "null"
          description: Average rating 0–5. Pass `null` to clear.
        ratingCount:
          oneOf:
            - type: integer
              minimum: 0
            - type: "null"
          description: Number of ratings. Pass `null` to clear.
        externalIds:
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/ExternalIdRef"
            - type: "null"
          description: >-
            Replaces all external id refs for this course. `null` = leave
            unchanged; `[]` = remove all; an array = full replacement.
        sourceUpdatedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
          description: Upstream source last-updated timestamp. Pass `null` to clear.
      example:
        title: Pragmatic Clean Architecture (2nd ed.)
        slug: pragmatic-clean-architecture-2nd-ed
        level: intermediate
        language: en
        instructorIds:
          - clxvins0000000000000000001
        tagIds:
          - clxvtag0000000000000000001
...